Major Navigation Refactor
class Navigation has been refactored to include all movement that happens in the file explorer, this is now also done within the current 'tab[n]'.
All explorer.utils. that refecenced a navigation element have been updated to reflect this change and now use the navigation class (tab[0].Navigation).

Added : nav.SetPath(Config.Path); to OpenFileExplorer() to set the path of the navigation to the current path of the file explorer.